requirements.txt
google-adk>=0.1.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
google-cloud-bigquery>=3.17.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
